Add upstream parsing interface
- abstract parsing code behind interface to allow additional
methods such as Kubernetes discovery.
